// frontend/src/lib/modules/openapi/model/create-pet-dto.ts
/**
 * Pet Store API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreatePetDto { 
    /**
     * The name of the pet
     */
    name: string;
    /**
     * The species of the pet
     */
    species: string;
    /**
     * The age of the pet in years
     */
    age: number;
    /**
     * Optional description of the pet
     */
    description?: string;
}